Move cargo watch functionality to separate crate
