+++ crates/ra_lsp_server/Cargo.toml
@@ -27,10 +27,9 @@ ra_project_model = { path = "../ra_project_model" }
 ra_prof = { path = "../ra_prof" }
 ra_vfs_glob = { path = "../ra_vfs_glob" }
 env_logger = { version = "0.7.1", default-features = false, features = ["humantime"] }
-cargo_metadata = "0.9.1"
+ra_cargo_watch = { path = "../ra_cargo_watch" }
 
 [dev-dependencies]
-insta = "0.12.0"
 tempfile = "3"
 test_utils = { path = "../test_utils" }
